Rename nifti and nrrd image builders
