+++ tests/test_nifti.py
@@ -5,7 +5,7 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-from panimg.image_builders.nifti import image_builder_nifti
+from panimg.image_builders.metaio_nifti import image_builder_nifti
 from panimg.models import ColorSpace
 from panimg.panimg import _build_files
 from tests import RESOURCE_PATH
